saa::lock

Struct Lock

pub struct Lock { /* private fields */ }

Lock is a low-level locking primitive for both synchronous and asynchronous operations.

The locking semantics are similar to RwLock, however, Lock only provides low-level locking and releasing methods, hence forcing the user to manage the scope of acquired locks and the resources to protect.

Implementations

impl Lock

pub const MAX_SHARED_OWNERS: usize = 62usize

Maximum number of shared owners.

pub fn is_free(&self, mo: Ordering) -> bool

Returns true if the lock is currently free.

Examples

use saa::Lock;
use std::sync::atomic::Ordering::Relaxed;

let lock = Lock::default();
assert!(lock.is_free(Relaxed));

lock.lock_sync();
assert!(!lock.is_free(Relaxed));

pub fn is_locked(&self, mo: Ordering) -> bool

Returns true if an exclusive lock is currently held.

Examples

use saa::Lock;
use std::sync::atomic::Ordering::Relaxed;

let lock = Lock::default();
assert!(!lock.is_locked(Relaxed));

lock.lock_sync();
assert!(lock.is_locked(Relaxed));
assert!(!lock.is_shared(Relaxed));

pub fn is_shared(&self, mo: Ordering) -> bool

Returns true if shared locks are currently held.

Examples

use saa::Lock;
use std::sync::atomic::Ordering::Relaxed;

let lock = Lock::default();

assert!(!lock.is_shared(Relaxed));

lock.share_sync();
assert!(lock.is_shared(Relaxed));
assert!(!lock.is_locked(Relaxed));

pub fn is_poisoned(&self, mo: Ordering) -> bool

Returns true if the Lock is poisoned.

Examples

use saa::Lock;
use std::sync::atomic::Ordering::Relaxed;

let lock = Lock::default();
assert!(!lock.is_poisoned(Relaxed));

lock.lock_sync();
assert!(lock.poison_lock());
assert!(lock.is_poisoned(Relaxed));

pub async fn lock_async(&self) -> bool

Acquires an exclusive lock asynchronously.

Returns false if the lock is poisoned.

Examples

use saa::Lock;
use std::sync::atomic::Ordering::Relaxed;

let lock = Lock::default();

async {
    lock.lock_async().await;
    assert!(lock.is_locked(Relaxed));
    assert!(!lock.is_shared(Relaxed))
};

pub fn lock_sync(&self) -> bool

Acquires an exclusive lock synchronously.

Returns false if the lock is poisoned.

Examples

use saa::Lock;
use std::sync::atomic::Ordering::Relaxed;

let lock = Lock::default();

lock.lock_sync();

assert!(lock.is_locked(Relaxed));
assert!(!lock.try_share());

pub async fn lock_async_with<F: FnOnce()>(&self, begin_wait: F) -> bool

Acquires an exclusive lock asynchronously with a wait callback.

Returns false if the lock is poisoned. The callback is invoked when the task starts waiting for a lock.

Examples

use saa::Lock;
use std::sync::atomic::Ordering::Relaxed;

let lock = Lock::default();

async {
    lock.lock_async().await;
    assert!(lock.is_locked(Relaxed));
    assert!(!lock.is_shared(Relaxed))
};

pub fn lock_sync_with<F: FnOnce()>(&self, begin_wait: F) -> bool

Acquires an exclusive lock synchronously with a wait callback.

Returns false if the lock is poisoned. The callback is invoked when the task starts waiting for a lock.

Examples

use saa::Lock;
use std::sync::atomic::Ordering::Relaxed;

let lock = Lock::default();

lock.lock_sync();

assert!(lock.is_locked(Relaxed));
assert!(!lock.try_share());

pub fn try_lock(&self) -> bool

Tries to acquire an exclusive lock.

Returns false if the lock was not free.

Examples

use saa::Lock;

let lock = Lock::default();

assert!(lock.try_lock());
assert!(!lock.try_share());
assert!(!lock.try_lock());

pub async fn share_async(&self) -> bool

Acquires a shared lock asynchronously.

Returns false if the lock is poisoned.

Examples

use saa::Lock;
use std::sync::atomic::Ordering::Relaxed;

let lock = Lock::default();

async {
    lock.share_async().await;
    assert!(!lock.is_locked(Relaxed));
    assert!(lock.is_shared(Relaxed))
};

pub fn share_sync(&self) -> bool

Acquires a shared lock synchronously.

Returns false if the lock is poisoned.

Examples

use saa::Lock;
use std::sync::atomic::Ordering::Relaxed;

let lock = Lock::default();

lock.share_sync();

assert!(lock.is_shared(Relaxed));
assert!(!lock.try_lock());

pub async fn share_async_with<F: FnOnce()>(&self, begin_wait: F) -> bool

Acquires a shared lock asynchronously with a wait callback.

Returns false if the lock is poisoned. The callback is invoked when the task starts waiting for a lock.

Examples

use saa::Lock;
use std::sync::atomic::Ordering::Relaxed;

let lock = Lock::default();

async {
    lock.share_async().await;
    assert!(!lock.is_locked(Relaxed));
    assert!(lock.is_shared(Relaxed))
};

pub fn share_sync_with<F: FnOnce()>(&self, begin_wait: F) -> bool

Acquires a shared lock synchronously with a wait callback.

Returns false if the lock is poisoned. The callback is invoked when the task starts waiting for a lock.

Examples

use saa::Lock;
use std::sync::atomic::Ordering::Relaxed;

let lock = Lock::default();

lock.share_sync();

assert!(lock.is_shared(Relaxed));
assert!(!lock.try_lock());

pub fn try_share(&self) -> bool

Tries to acquire a shared lock.

Returns false if an exclusive lock is held, the number of shared owners has reached Self::MAX_SHARED_OWNERS, or the lock is poisoned.

Examples

use saa::Lock;

let lock = Lock::default();

assert!(lock.try_share());
assert!(lock.try_share());
assert!(!lock.try_lock());

pub fn register_pager<'l>( &'l self, pager: &mut Pin<&mut Pager<'l, Self>>, mode: Mode, is_sync: bool, ) -> bool

Registers a Pager to allow it to get an exclusive lock or a shared lock remotely.

is_sync indicates whether the Pager will be polled asynchronously (false) or synchronously (true).

Returns false if the Pager was already registered.

Examples

use std::pin::pin;

use saa::{Lock, Pager};
use saa::lock::Mode;

let lock = Lock::default();

let mut pinned_pager = pin!(Pager::default());

assert!(lock.register_pager(&mut pinned_pager, Mode::Exclusive, true));
assert!(!lock.register_pager(&mut pinned_pager, Mode::Exclusive, true));

assert_eq!(pinned_pager.poll_sync(), Ok(true));

pub fn release_lock(&self) -> bool

Releases an exclusive lock.

Returns true if an exclusive lock was previously held and successfully released.

Examples

use saa::Lock;

let lock = Lock::default();

lock.lock_sync();

assert!(lock.release_lock());
assert!(!lock.release_lock());

assert!(lock.try_share());
assert!(!lock.release_lock());
assert!(lock.release_share());

lock.lock_sync();
lock.poison_lock();

assert!(!lock.release_lock());

pub fn poison_lock(&self) -> bool

Poisons the lock with an exclusive lock held.

Returns false if an exclusive lock is not held, or the lock was already poisoned.

Examples

use std::sync::atomic::Ordering::Relaxed;

use saa::Lock;

let lock = Lock::default();

assert!(!lock.poison_lock());

lock.lock_sync();

assert!(lock.poison_lock());
assert!(lock.is_poisoned(Relaxed));

assert!(!lock.poison_lock());

assert!(!lock.lock_sync());
assert!(!lock.share_sync());
assert!(!lock.release_lock());
assert!(!lock.release_share());

pub fn clear_poison(&self) -> bool

Clears poison from the lock.

Returns true if the lock was successfully cleared.

Examples

use std::sync::atomic::Ordering::Relaxed;

use saa::Lock;

let lock = Lock::default();

assert!(!lock.poison_lock());

lock.lock_sync();

assert!(lock.poison_lock());
assert!(lock.clear_poison());
assert!(!lock.is_poisoned(Relaxed));

pub fn release_share(&self) -> bool

Releases a shared lock.

Returns true if a shared lock was previously held and successfully released.

Examples

use saa::Lock;

let lock = Lock::default();

lock.share_sync();
lock.share_sync();

assert!(lock.release_share());

assert!(!lock.try_lock());
assert!(lock.release_share());

assert!(!lock.release_share());
assert!(lock.try_lock());

lock.poison_lock();

assert!(!lock.release_share());