saa::semaphore

Struct Semaphore

pub struct Semaphore { /* private fields */ }

Semaphore is a synchronization primitive that allows a fixed number of threads to access a resource concurrently.

Implementations

impl Semaphore

pub const MAX_PERMITS: usize = 63usize

Maximum number of concurrent owners.

pub fn with_permits(permits: usize) -> Self

Creates a new Semaphore with the given number of initially available permits.

The maximum number of available permits is MAX_PERMITS, and if a value greater than or equal to MAX_PERMITS is provided, it will be set to MAX_PERMITS.

Examples

use saa::Semaphore;
use std::sync::atomic::Ordering::Relaxed;

let semaphore = Semaphore::with_permits(11);

assert_eq!(semaphore.available_permits(Relaxed), 11);

assert!(semaphore.try_acquire_many(11));
assert!(!semaphore.is_open(Relaxed));

pub fn is_open(&self, mo: Ordering) -> bool

Returns true if the semaphore is currently open.

Examples

use saa::Semaphore;
use std::sync::atomic::Ordering::Relaxed;

let semaphore = Semaphore::default();
assert!(semaphore.is_open(Relaxed));

assert!(semaphore.try_acquire_many(Semaphore::MAX_PERMITS));
assert!(!semaphore.is_open(Relaxed));

pub fn is_closed(&self, mo: Ordering) -> bool

Returns true if the semaphore is currently closed.

Examples

use saa::Semaphore;
use std::sync::atomic::Ordering::Relaxed;

let semaphore = Semaphore::default();
assert!(!semaphore.is_closed(Relaxed));
assert!(semaphore.is_open(Relaxed));

assert!(semaphore.try_acquire());
assert!(!semaphore.is_closed(Relaxed));
assert!(semaphore.is_open(Relaxed));

semaphore.try_acquire_many(Semaphore::MAX_PERMITS - 1);
assert!(semaphore.is_closed(Relaxed));

pub fn available_permits(&self, mo: Ordering) -> usize

Returns the number of available permits.

Examples

use saa::Semaphore;
use std::sync::atomic::Ordering::Relaxed;

let semaphore = Semaphore::default();
assert_eq!(semaphore.available_permits(Relaxed), Semaphore::MAX_PERMITS);

assert!(semaphore.try_acquire());
assert_eq!(semaphore.available_permits(Relaxed), Semaphore::MAX_PERMITS - 1);

pub async fn acquire_async(&self)

Gets a permit from the semaphore asynchronously.

Examples

use saa::Semaphore;
use std::sync::atomic::Ordering::Relaxed;

let semaphore = Semaphore::default();

async {
    semaphore.acquire_async().await;
    assert_eq!(semaphore.available_permits(Relaxed), Semaphore::MAX_PERMITS - 1);
};

pub fn acquire_sync(&self)

Gets a permit from the semaphore synchronously.

Examples

use saa::Semaphore;
use std::sync::atomic::Ordering::Relaxed;

let semaphore = Semaphore::default();

semaphore.acquire_sync();
assert_eq!(semaphore.available_permits(Relaxed), Semaphore::MAX_PERMITS - 1);

pub async fn acquire_async_with<F: FnOnce()>(&self, begin_wait: F)

Gets a permit from the semaphore asynchronously with a wait callback.

The callback is invoked when the task starts waiting for a permit.

Examples

use saa::Semaphore;
use std::sync::atomic::Ordering::Relaxed;

let semaphore = Semaphore::default();

async {
    let mut wait = false;
    semaphore.acquire_async_with(|| { wait = true; }).await;
    assert_eq!(semaphore.available_permits(Relaxed), Semaphore::MAX_PERMITS - 1);
    assert!(!wait);
};

pub fn acquire_sync_with<F: FnOnce()>(&self, begin_wait: F)

Gets multiple permits from the semaphore synchronously with a wait callback.

The callback is invoked when the task starts waiting for permits.

Examples

use saa::Semaphore;
use std::sync::atomic::Ordering::Relaxed;

let semaphore = Semaphore::default();

let mut wait = false;
semaphore.acquire_sync_with(|| { wait = true; });
assert_eq!(semaphore.available_permits(Relaxed), Semaphore::MAX_PERMITS - 1);
assert!(!wait);

pub fn try_acquire(&self) -> bool

Tries to get a permit from the semaphore.

Returns false if no permits are available.

Examples

use saa::Semaphore;

let semaphore = Semaphore::default();

assert!(semaphore.try_acquire());
assert!(!semaphore.try_acquire_many(Semaphore::MAX_PERMITS));

pub async fn acquire_many_async(&self, count: usize) -> bool

Gets multiple permits from the semaphore asynchronously.

Returns false if the count exceeds Self::MAX_PERMITS.

Examples

use saa::Semaphore;
use std::sync::atomic::Ordering::Relaxed;

let semaphore = Semaphore::default();

async {
    assert!(semaphore.acquire_many_async(11).await);
    assert_eq!(semaphore.available_permits(Relaxed), Semaphore::MAX_PERMITS - 11);
};

pub fn acquire_many_sync(&self, count: usize) -> bool

Gets multiple permits from the semaphore synchronously.

Returns false if the count exceeds Self::MAX_PERMITS.

Examples

use saa::Semaphore;
use std::sync::atomic::Ordering::Relaxed;

let semaphore = Semaphore::default();

assert!(semaphore.acquire_many_sync(11));
assert_eq!(semaphore.available_permits(Relaxed), Semaphore::MAX_PERMITS - 11);

pub async fn acquire_many_async_with<F: FnOnce()>( &self, count: usize, begin_wait: F, ) -> bool

Gets multiple permits from the semaphore asynchronously with a wait callback.

Returns false if the count exceeds Self::MAX_PERMITS. The callback is invoked when the task starts waiting for permits.

Examples

use saa::Semaphore;
use std::sync::atomic::Ordering::Relaxed;

let semaphore = Semaphore::default();

async {
    let mut wait = false;
    assert!(semaphore.acquire_many_async_with(2, || { wait = true; }).await);
    assert_eq!(semaphore.available_permits(Relaxed), Semaphore::MAX_PERMITS - 2);
    assert!(!wait);
};

pub fn acquire_many_sync_with<F: FnOnce()>( &self, count: usize, begin_wait: F, ) -> bool

Gets multiple permits from the semaphore synchronously with a wait callback.

Returns false if the count exceeds Self::MAX_PERMITS. The callback is invoked when the task starts waiting for permits.

Examples

use saa::Semaphore;
use std::sync::atomic::Ordering::Relaxed;

let semaphore = Semaphore::default();

let mut wait = false;
assert!(semaphore.acquire_many_sync_with(2, || { wait = true; }));
assert_eq!(semaphore.available_permits(Relaxed), Semaphore::MAX_PERMITS - 2);
assert!(!wait);

pub fn try_acquire_many(&self, count: usize) -> bool

Tries to get multiple permits from the semaphore.

Returns false if no permits are available.

Examples

use saa::Semaphore;

let semaphore = Semaphore::default();

assert!(semaphore.try_acquire_many(Semaphore::MAX_PERMITS));
assert!(!semaphore.try_acquire());

pub fn register_pager<'s>( &'s self, pager: &mut Pin<&mut Pager<'s, Self>>, count: usize, is_sync: bool, ) -> bool

Registers a Pager to allow it to get a permit remotely.

is_sync indicates whether the Pager will be polled asynchronously (false) or synchronously (true).

Returns false if the Pager was already registered, or if the count is greater than the maximum number of permits.

Examples

use std::pin::pin;

use saa::{Pager, Semaphore};

let semaphore = Semaphore::default();

let mut pinned_pager = pin!(Pager::default());

assert!(semaphore.register_pager(&mut pinned_pager, 1, true));
assert!(!semaphore.register_pager(&mut pinned_pager, 1, true));

assert!(pinned_pager.poll_sync().is_ok());

pub fn release(&self) -> bool

Releases a permit.

Returns true if a permit was successfully released.

Examples

use saa::Semaphore;
use std::sync::atomic::Ordering::Relaxed;

let semaphore = Semaphore::default();

assert!(semaphore.try_acquire_many(11));
assert_eq!(semaphore.available_permits(Relaxed), Semaphore::MAX_PERMITS - 11);

assert!(semaphore.release());
assert_eq!(semaphore.available_permits(Relaxed), Semaphore::MAX_PERMITS - 10);

pub fn release_many(&self, count: usize) -> bool

Releases permits.

Returns true if the specified number of permits were successfully released.

Examples

use saa::Semaphore;
use std::sync::atomic::Ordering::Relaxed;

let semaphore = Semaphore::default();

assert!(semaphore.try_acquire_many(11));
assert_eq!(semaphore.available_permits(Relaxed), Semaphore::MAX_PERMITS - 11);

assert!(semaphore.release_many(10));
assert_eq!(semaphore.available_permits(Relaxed), Semaphore::MAX_PERMITS - 1);